pyproject.toml
[tool.poetry]
name = "pdftimetableparser"
version = "0.1.0"
description = ""
authors = ["Ihor Savenko <csigorek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
camelot-py = "^0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
